pyproject.toml
[project]
name = "tinkoff-invest-mcp"
version = "0.1.0"
description = "Tinkoff Invest MCP Server"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastmcp>=0.1.0",
    "tinkoff-investments>=0.2.0b116",
    "pydantic>=2.0.0",
]

[project.scripts]
tinkoff-invest-mcp = "tinkoff_invest_mcp.server:main"
tinkoff-sandbox = "tinkoff_invest_mcp.cli:cli_main"

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.3.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.12",
    # Stub packages для mypy
    "types-requests>=2.0.0",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "PTH",  # flake8-use-pathlib
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "B904",  # use raise from
    "RUF001",  # ambiguous unicode character (кириллица в строках)
    "RUF002",  # ambiguous unicode character (кириллица в docstring)
    "RUF003",  # ambiguous unicode character (кириллица в комментариях)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
extra_checks = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
    "fastmcp.*",
    "tinkoff.*",
    "dotenv.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
testpaths = ["tests"]

[tool.coverage.run]
omit = [
    "src/tinkoff_invest_mcp/cli.py",
]
